requirements.txt
Flask==3.0.1
playwright==1.41.0
dnspython==2.5.0